Separate out the handlers
* make them more modular and write tests
