+++ server/version.go
@@ -1,4 +1,4 @@
 package server
 
 // Version exported
-const Version string = "0.2.2"
+const Version string = "0.2.3"
